+++ NEWS.md
@@ -1,5 +1,7 @@
 # phsopendata (development version)
 
+- `get_dataset()` can now resolve a human-readable dataset title to its dataset name, with a warning. (#80)
+
 # phsopendata 1.1.0 (2026-06-05)
 
 - [`list_resources()`](https://public-health-scotland.github.io/phsopendata/reference/list_resources.html) has been upgraded. Instead of just returning all resources from a dataset (given the exact title), it can now search both resources and dataset titles and will list all resources that match the search!